Improve gamepad binding UI
