+++ src/renderer/components/pages/Config/ConfigPage.tsx
@@ -1,16 +1,12 @@
 import React, { FC } from 'react'
-import {
-  ConfigPageGrid,
-  MenuArea,
-  ConfigRouterArea,
-} from '@/renderer/components/pages/Config/ConfigPage.styles'
 import { Route, Switch, Redirect } from 'react-router'
 import { ConfigMenu } from './ConfigMenu'
 import { GeneralConfig } from '@/renderer/components/pages/Config/pages/GeneralConfig'
 import { CameraConfig } from '@/renderer/components/pages/Config/pages/CameraConfig/CameraConfig'
-import { GamepadConfig } from '@/renderer/components/pages/Config/pages/GamepadConfig/GamepadConfig'
+import { GamepadConfig } from '@/renderer/components/pages/Config/pages/GamepadConfig'
 import { useControl } from '@/renderer/utils/hooks/useControl'
 import { GraphConfig } from '@/renderer/components/pages/Config/pages/GraphConfig/GraphConfig'
+import { styled } from '@/renderer/globalStyles/styled'
 
 export const ConfigPage: FC = () => {
   useControl('nothing')
@@ -31,3 +27,24 @@ export const ConfigPage: FC = () => {
     </ConfigPageGrid>
   )
 }
+
+const ConfigPageGrid = styled.div`
+  display: grid;
+  grid-template: 'm r';
+  grid-template-columns: 1fr 5fr;
+  height: 100%;
+  max-height: 100%;
+`
+
+const MenuArea = styled.div`
+  grid-area: m;
+  background-color: ${({ theme }) => theme.colors.darkerBackground};
+  margin: 0;
+  min-height: 100%;
+`
+
+const ConfigRouterArea = styled.div`
+  grid-area: r;
+  padding: 16px;
+  overflow: auto;
+`
