--- src/renderer/components/pages/Config/pages/CameraConfig/Table.styles.ts
@@ -1,42 +1,0 @@
-import { styled } from '@/renderer/globalStyles/styled'
-
-export const StyledTable = styled.table`
-  width: 100%;
-  border-spacing: 0;
-  border-collapse: collapse;
-
-  th,
-  td {
-    text-align: left;
-
-    &:last-child {
-      width: 32px;
-    }
-  }
-
-  td {
-    padding: 8px 8px;
-
-    &:last-child {
-      cursor: pointer;
-    }
-
-    input {
-      width: 100%;
-    }
-
-    select {
-      width: 100%;
-    }
-  }
-
-  thead th {
-    padding: 4px 8px;
-    font-size: inherit;
-    border-bottom: 1px solid ${({ theme }) => theme.colors.border};
-  }
-
-  tbody {
-    border: 1px solid ${({ theme }) => theme.colors.border};
-  }
-`
