+++ src/renderer/components/pages/Config/pages/CameraConfig/Table.tsx
@@ -11,7 +11,6 @@ import {
 } from '@/renderer/store/modules/feed/@types'
 import { FaTimes } from 'react-icons/fa'
 import { useSelector } from '@/renderer/utils/hooks/typedUseSelector'
-import { StyledTable } from './Table.styles'
 import { Button } from '@/renderer/components/common/Button'
 import { Modal } from '@/renderer/components/common/Modal/Modal'
 import { useOpenClose } from '@/renderer/utils/hooks/useOpenClose'
@@ -20,6 +19,7 @@ import { rosService, videoUrlSelector } from '@/renderer/state/ros'
 import { Select } from '@/renderer/components/common/Select'
 import { Input } from '@/renderer/components/common/Input'
 import { useEscape } from '@/renderer/utils/hooks/useEscape'
+import { styled } from '@/renderer/globalStyles/styled'
 
 interface TableRowProps {
   feed: ICameraFeed
@@ -143,3 +143,44 @@ export const Table: FC = () => {
     </StyledTable>
   )
 }
+
+const StyledTable = styled.table`
+  width: 100%;
+  border-spacing: 0;
+  border-collapse: collapse;
+
+  th,
+  td {
+    text-align: left;
+
+    &:last-child {
+      width: 32px;
+    }
+  }
+
+  td {
+    padding: 8px 8px;
+
+    &:last-child {
+      cursor: pointer;
+    }
+
+    input {
+      width: 100%;
+    }
+
+    select {
+      width: 100%;
+    }
+  }
+
+  thead th {
+    padding: 4px 8px;
+    font-size: inherit;
+    border-bottom: 1px solid ${({ theme }) => theme.colors.border};
+  }
+
+  tbody {
+    border: 1px solid ${({ theme }) => theme.colors.border};
+  }
+`
